Move HandleAuth component to auth folder
